[qt] Fix party_type eventing: dedupe Qt controller, move to RefdataPlugin
party_type's Qt code existed as two full duplicates: a hand-written
copy in ores.qt/party (still wired up by PartyPlugin, but constructed
with an empty event name so it silently never subscribed to NATS
eventing) and a correctly codegen-generated copy in ores.qt/refdata
(never compiled into any running plugin). The old copy also lacked the
ChangeReasonCache parameter present on its sibling PartyStatusController
— a feature regression from an earlier, incomplete Qt codegen
migration.

Root cause: every entity in ores.qt.party has backend component=refdata;
there is no ores.party C++ component. The plugin is a UI-only grouping
with zero backend justification, and that mismatch is what let the
duplication go unnoticed.

Fixed by:
- Adding :has_change_reason_cache: true to party_type's Qt model
  properties (matching currency/party_status), regenerating its Qt
  profile — codegen already places it under ores.qt/refdata by
  default, matching its backend component, no override needed.
- Deleting the stale ores.qt/party duplicate.
- Moving PartyTypeController instantiation from PartyPlugin to
  RefdataPlugin.
- Adding shared_menus_context::organisation_codes_menu (host-owned in
  MainWindow.cpp) so RefdataPlugin and PartyPlugin can both contribute
  to the same physical "Organisation Codes" submenu without a
  cross-plugin library dependency, since the remaining ores.qt.party
  entities migrate one at a time as each is (re-)commissioned.

Verified: full ores.qt.exe builds clean; live end-to-end retest (fresh
Barclays Plc tenant, full 24-service fleet, Qt client login) confirms
party_type list/detail/save/eventing all work correctly.

Task-ID: 21E5D45B-9CAD-4146-8ECF-1ACDB6398508
