--- projects/ores.qt/party/include/ores.qt/PartyTypeHistoryDialog.hpp
@@ -1,92 +1,0 @@
-/* -*- mode: c++; tab-width: 4; indent-tabs-mode: nil; c-basic-offset: 4 -*-
- *
- * Copyright (C) 2026 Marco Craveiro <marco.craveiro@gmail.com>
- *
- * This program is free software; you can redistribute it and/or modify it under
- * the terms of the GNU General Public License as published by the Free Software
- * Foundation; either version 3 of the License, or (at your option) any later
- * version.
- *
- * This program is distributed in the hope that it will be useful, but WITHOUT
- * ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or FITNESS
- * FOR A PARTICULAR PURPOSE. See the GNU General Public License for more
- * details.
- *
- * You should have received a copy of the GNU General Public License along with
- * this program; if not, write to the Free Software Foundation, Inc., 51
- * Franklin Street, Fifth Floor, Boston, MA 02110-1301, USA.
- *
- */
-#ifndef ORES_QT_PARTY_TYPE_HISTORY_DIALOG_HPP
-#define ORES_QT_PARTY_TYPE_HISTORY_DIALOG_HPP
-
-#include "ores.logging/make_logger.hpp"
-#include "ores.qt/ClientManager.hpp"
-#include "ores.qt/HistoryDialogBase.hpp"
-#include "ores.refdata.api/domain/party_type.hpp"
-#include <QString>
-#include <memory>
-#include <vector>
-
-namespace Ui {
-class PartyTypeHistoryDialog;
-}
-
-namespace ores::qt {
-
-/**
- * @brief Dialog for viewing the version history of a party type.
- *
- * Shows all historical versions of a party type with ability
- * to view details or revert to a previous version.
- */
-class PartyTypeHistoryDialog final : public HistoryDialogBase {
-    Q_OBJECT
-
-private:
-    inline static std::string_view logger_name = "ores.qt.party_type_history_dialog";
-
-    [[nodiscard]] static auto& lg() {
-        using namespace ores::logging;
-        static auto instance = make_logger(logger_name);
-        return instance;
-    }
-
-public:
-    explicit PartyTypeHistoryDialog(const QString& code,
-                                    ClientManager* clientManager,
-                                    QWidget* parent = nullptr);
-    ~PartyTypeHistoryDialog() override;
-
-    void loadHistory() override;
-
-    /**
-     * @brief Returns the identifier of the party type.
-     */
-    [[nodiscard]] QString code() const override {
-        return code_;
-    }
-
-signals:
-    void openVersionRequested(const refdata::domain::party_type& type, int versionNumber);
-    void revertVersionRequested(const refdata::domain::party_type& type);
-
-protected:
-    [[nodiscard]] int historySize() const override;
-    [[nodiscard]] VersionRow versionRow(int index) const override;
-    [[nodiscard]] QString historyTitle() const override;
-    [[nodiscard]] DiffResult calculateDiffAt(int current_index, int previous_index) const override;
-    void displayFullDetails(int index) override;
-    void openVersionAt(int index) override;
-    void revertToVersionAt(int index) override;
-
-private:
-    std::unique_ptr<Ui::PartyTypeHistoryDialog> ui_;
-    QString code_;
-    ClientManager* clientManager_;
-    std::vector<refdata::domain::party_type> versions_;
-};
-
-}
-
-#endif
