--- projects/ores.qt/party/src/PartyTypeController.cpp
@@ -1,433 +1,0 @@
-/* -*- mode: c++; tab-width: 4; indent-tabs-mode: nil; c-basic-offset: 4 -*-
- *
- * Copyright (C) 2026 Marco Craveiro <marco.craveiro@gmail.com>
- *
- * This program is free software; you can redistribute it and/or modify it under
- * the terms of the GNU General Public License as published by the Free Software
- * Foundation; either version 3 of the License, or (at your option) any later
- * version.
- *
- * This program is distributed in the hope that it will be useful, but WITHOUT
- * ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or FITNESS
- * FOR A PARTICULAR PURPOSE. See the GNU General Public License for more
- * details.
- *
- * You should have received a copy of the GNU General Public License along with
- * this program; if not, write to the Free Software Foundation, Inc., 51
- * Franklin Street, Fifth Floor, Boston, MA 02110-1301, USA.
- *
- */
-#include "ores.qt/PartyTypeController.hpp"
-#include "ores.qt/ChangeReasonCache.hpp"
-#include "ores.qt/DetachableMdiSubWindow.hpp"
-#include "ores.qt/IconUtils.hpp"
-#include "ores.qt/PartyTypeDetailDialog.hpp"
-#include "ores.qt/PartyTypeHistoryDialog.hpp"
-#include "ores.qt/PartyTypeMdiWindow.hpp"
-#include <QMdiSubWindow>
-#include <QMessageBox>
-#include <QPointer>
-
-namespace ores::qt {
-
-using namespace ores::logging;
-
-PartyTypeController::PartyTypeController(QMainWindow* mainWindow,
-                                         QMdiArea* mdiArea,
-                                         ClientManager* clientManager,
-                                         ChangeReasonCache* changeReasonCache,
-                                         const QString& username,
-                                         QObject* parent)
-    : EntityController(mainWindow, mdiArea, clientManager, username, std::string_view{}, parent)
-    , changeReasonCache_(changeReasonCache)
-    , listWindow_(nullptr)
-    , listMdiSubWindow_(nullptr) {
-
-    BOOST_LOG_SEV(lg(), debug) << "PartyTypeController created";
-}
-
-void PartyTypeController::showListWindow() {
-    BOOST_LOG_SEV(lg(), debug) << "showListWindow called";
-
-    const QString key = build_window_key("list", "party_types");
-    if (try_reuse_window(key)) {
-        BOOST_LOG_SEV(lg(), debug) << "Reusing existing list window";
-        return;
-    }
-
-    // Create new window
-    listWindow_ = new PartyTypeMdiWindow(clientManager_, username_);
-
-    // Connect signals
-    connect(
-        listWindow_, &PartyTypeMdiWindow::statusChanged, this, &PartyTypeController::statusMessage);
-    connect(
-        listWindow_, &PartyTypeMdiWindow::errorOccurred, this, &PartyTypeController::errorMessage);
-    connect(listWindow_,
-            &PartyTypeMdiWindow::showTypeDetails,
-            this,
-            &PartyTypeController::onShowDetails);
-    connect(listWindow_,
-            &PartyTypeMdiWindow::addNewRequested,
-            this,
-            &PartyTypeController::onAddNewRequested);
-    connect(listWindow_,
-            &PartyTypeMdiWindow::showTypeHistory,
-            this,
-            &PartyTypeController::onShowHistory);
-
-    // Create MDI subwindow
-    listMdiSubWindow_ = new DetachableMdiSubWindow(mainWindow_);
-    listMdiSubWindow_->setWidget(listWindow_);
-    listMdiSubWindow_->setWindowTitle("Party Types");
-    listMdiSubWindow_->setWindowIcon(
-        IconUtils::createRecoloredIcon(Icon::Tag, IconUtils::DefaultIconColor));
-    listMdiSubWindow_->setAttribute(Qt::WA_DeleteOnClose);
-    listMdiSubWindow_->resize(listWindow_->sizeHint());
-
-    mdiArea_->addSubWindow(listMdiSubWindow_);
-    listMdiSubWindow_->show();
-
-    // Track window
-    track_window(key, listMdiSubWindow_);
-    register_detachable_window(listMdiSubWindow_);
-
-    // Cleanup when closed
-    connect(listMdiSubWindow_,
-            &QObject::destroyed,
-            this,
-            [self = QPointer<PartyTypeController>(this), key]() {
-                if (!self)
-                    return;
-                self->untrack_window(key);
-                self->listWindow_ = nullptr;
-                self->listMdiSubWindow_ = nullptr;
-            });
-
-    BOOST_LOG_SEV(lg(), debug) << "Party Type list window created";
-}
-
-void PartyTypeController::closeAllWindows() {
-    BOOST_LOG_SEV(lg(), debug) << "closeAllWindows called";
-
-    // Close all managed windows
-    QList<QString> keys = managed_windows_.keys();
-    for (const QString& key : keys) {
-        if (auto* window = managed_windows_.value(key)) {
-            window->close();
-        }
-    }
-    managed_windows_.clear();
-
-    listWindow_ = nullptr;
-    listMdiSubWindow_ = nullptr;
-}
-
-void PartyTypeController::reloadListWindow() {
-    if (listWindow_) {
-        listWindow_->reload();
-    }
-}
-
-void PartyTypeController::onShowDetails(const refdata::domain::party_type& type) {
-    BOOST_LOG_SEV(lg(), debug) << "Show details for: " << type.code;
-    showDetailWindow(type);
-}
-
-void PartyTypeController::onAddNewRequested() {
-    BOOST_LOG_SEV(lg(), info) << "Add new party type requested";
-    showAddWindow();
-}
-
-void PartyTypeController::onShowHistory(const refdata::domain::party_type& type) {
-    BOOST_LOG_SEV(lg(), debug) << "Show history requested for: " << type.code;
-    showHistoryWindow(QString::fromStdString(type.code));
-}
-
-void PartyTypeController::showAddWindow() {
-    BOOST_LOG_SEV(lg(), debug) << "Creating add window for new party type";
-
-    auto* detailDialog = new PartyTypeDetailDialog(mainWindow_);
-    if (changeReasonCache_)
-        detailDialog->setChangeReasonCache(changeReasonCache_);
-    detailDialog->setClientManager(clientManager_);
-    detailDialog->setUsername(username_.toStdString());
-    detailDialog->setCreateMode(true);
-
-    connect(detailDialog,
-            &PartyTypeDetailDialog::statusMessage,
-            this,
-            &PartyTypeController::statusMessage);
-    connect(detailDialog,
-            &PartyTypeDetailDialog::errorMessage,
-            this,
-            &PartyTypeController::errorMessage);
-    connect(detailDialog,
-            &PartyTypeDetailDialog::typeSaved,
-            this,
-            [self = QPointer<PartyTypeController>(this)](const QString& code) {
-                if (!self)
-                    return;
-                BOOST_LOG_SEV(lg(), info) << "Party Type saved: " << code.toStdString();
-                self->handleEntitySaved();
-            });
-
-    auto* detailWindow = new DetachableMdiSubWindow(mainWindow_);
-    detailWindow->setAttribute(Qt::WA_DeleteOnClose);
-    detailWindow->setWidget(detailDialog);
-    detailWindow->setWindowTitle("New Party Type");
-    detailWindow->setWindowIcon(
-        IconUtils::createRecoloredIcon(Icon::Tag, IconUtils::DefaultIconColor));
-
-    register_detachable_window(detailWindow);
-
-    connect_dialog_close(detailDialog, detailWindow);
-    show_managed_window(detailWindow, listMdiSubWindow_);
-}
-
-void PartyTypeController::showDetailWindow(const refdata::domain::party_type& type) {
-
-    const QString identifier = QString::fromStdString(type.code);
-    const QString key = build_window_key("details", identifier);
-
-    if (try_reuse_window(key)) {
-        BOOST_LOG_SEV(lg(), debug) << "Reusing existing detail window";
-        return;
-    }
-
-    BOOST_LOG_SEV(lg(), debug) << "Creating detail window for: " << type.code;
-
-    auto* detailDialog = new PartyTypeDetailDialog(mainWindow_);
-    if (changeReasonCache_)
-        detailDialog->setChangeReasonCache(changeReasonCache_);
-    detailDialog->setClientManager(clientManager_);
-    detailDialog->setUsername(username_.toStdString());
-    detailDialog->setCreateMode(false);
-    detailDialog->setType(type);
-
-    connect(detailDialog,
-            &PartyTypeDetailDialog::statusMessage,
-            this,
-            &PartyTypeController::statusMessage);
-    connect(detailDialog,
-            &PartyTypeDetailDialog::errorMessage,
-            this,
-            &PartyTypeController::errorMessage);
-    connect(detailDialog,
-            &PartyTypeDetailDialog::typeSaved,
-            this,
-            [self = QPointer<PartyTypeController>(this)](const QString& code) {
-                if (!self)
-                    return;
-                BOOST_LOG_SEV(lg(), info) << "Party Type saved: " << code.toStdString();
-                self->handleEntitySaved();
-            });
-    connect(detailDialog,
-            &PartyTypeDetailDialog::typeDeleted,
-            this,
-            [self = QPointer<PartyTypeController>(this), key](const QString& code) {
-                if (!self)
-                    return;
-                BOOST_LOG_SEV(lg(), info) << "Party Type deleted: " << code.toStdString();
-                self->handleEntityDeleted();
-            });
-
-    auto* detailWindow = new DetachableMdiSubWindow(mainWindow_);
-    detailWindow->setAttribute(Qt::WA_DeleteOnClose);
-    detailWindow->setWidget(detailDialog);
-    detailWindow->setWindowTitle(QString("Party Type: %1").arg(identifier));
-    detailWindow->setWindowIcon(
-        IconUtils::createRecoloredIcon(Icon::Tag, IconUtils::DefaultIconColor));
-
-    // Track window
-    track_window(key, detailWindow);
-    register_detachable_window(detailWindow);
-
-    QPointer<PartyTypeController> self = this;
-    connect(detailWindow, &QObject::destroyed, this, [self, key]() {
-        if (self) {
-            self->untrack_window(key);
-        }
-    });
-
-    connect_dialog_close(detailDialog, detailWindow);
-    show_managed_window(detailWindow, listMdiSubWindow_);
-}
-
-void PartyTypeController::showHistoryWindow(const QString& code) {
-    BOOST_LOG_SEV(lg(), info) << "Opening history window for party type: " << code.toStdString();
-
-    const QString windowKey = build_window_key("history", code);
-
-    // Try to reuse existing window
-    if (try_reuse_window(windowKey)) {
-        BOOST_LOG_SEV(lg(), info) << "Reusing existing history window for: " << code.toStdString();
-        return;
-    }
-
-    BOOST_LOG_SEV(lg(), info) << "Creating new history window for: " << code.toStdString();
-
-    auto* historyDialog = new PartyTypeHistoryDialog(code, clientManager_, mainWindow_);
-
-    connect(historyDialog,
-            &PartyTypeHistoryDialog::statusChanged,
-            this,
-            [self = QPointer<PartyTypeController>(this)](const QString& message) {
-                if (!self)
-                    return;
-                emit self->statusMessage(message);
-            });
-    connect(historyDialog,
-            &PartyTypeHistoryDialog::errorOccurred,
-            this,
-            [self = QPointer<PartyTypeController>(this)](const QString& message) {
-                if (!self)
-                    return;
-                emit self->errorMessage(message);
-            });
-    connect(historyDialog,
-            &PartyTypeHistoryDialog::revertVersionRequested,
-            this,
-            &PartyTypeController::onRevertVersion);
-    connect(historyDialog,
-            &PartyTypeHistoryDialog::openVersionRequested,
-            this,
-            &PartyTypeController::onOpenVersion);
-
-    // Load history data
-    historyDialog->loadHistory();
-
-    auto* historyWindow = new DetachableMdiSubWindow(mainWindow_);
-    historyWindow->setAttribute(Qt::WA_DeleteOnClose);
-    historyWindow->setWidget(historyDialog);
-    historyWindow->setWindowTitle(QString("Party Type History: %1").arg(code));
-    historyWindow->setWindowIcon(
-        IconUtils::createRecoloredIcon(Icon::History, IconUtils::DefaultIconColor));
-
-    // Track this history window
-    track_window(windowKey, historyWindow);
-    register_detachable_window(historyWindow);
-
-    QPointer<PartyTypeController> self = this;
-    connect(historyWindow, &QObject::destroyed, this, [self, windowKey]() {
-        if (self) {
-            self->untrack_window(windowKey);
-        }
-    });
-
-    show_managed_window(historyWindow, listMdiSubWindow_);
-}
-
-void PartyTypeController::onOpenVersion(const refdata::domain::party_type& type,
-                                        int versionNumber) {
-    BOOST_LOG_SEV(lg(), info) << "Opening historical version " << versionNumber
-                              << " for party type: " << type.code;
-
-    const QString code = QString::fromStdString(type.code);
-    const QString windowKey =
-        build_window_key("version", QString("%1_v%2").arg(code).arg(versionNumber));
-
-    // Try to reuse existing window
-    if (try_reuse_window(windowKey)) {
-        BOOST_LOG_SEV(lg(), info) << "Reusing existing version window";
-        return;
-    }
-
-    auto* detailDialog = new PartyTypeDetailDialog(mainWindow_);
-    if (changeReasonCache_)
-        detailDialog->setChangeReasonCache(changeReasonCache_);
-    detailDialog->setClientManager(clientManager_);
-    detailDialog->setUsername(username_.toStdString());
-    detailDialog->setType(type);
-    detailDialog->setReadOnly(true);
-
-    connect(detailDialog,
-            &PartyTypeDetailDialog::statusMessage,
-            this,
-            [self = QPointer<PartyTypeController>(this)](const QString& message) {
-                if (!self)
-                    return;
-                emit self->statusMessage(message);
-            });
-    connect(detailDialog,
-            &PartyTypeDetailDialog::errorMessage,
-            this,
-            [self = QPointer<PartyTypeController>(this)](const QString& message) {
-                if (!self)
-                    return;
-                emit self->errorMessage(message);
-            });
-
-    auto* detailWindow = new DetachableMdiSubWindow(mainWindow_);
-    detailWindow->setAttribute(Qt::WA_DeleteOnClose);
-    detailWindow->setWidget(detailDialog);
-    detailWindow->setWindowTitle(
-        QString("Party Type: %1 (Version %2)").arg(code).arg(versionNumber));
-    detailWindow->setWindowIcon(
-        IconUtils::createRecoloredIcon(Icon::History, IconUtils::DefaultIconColor));
-
-    track_window(windowKey, detailWindow);
-    register_detachable_window(detailWindow);
-
-    QPointer<PartyTypeController> self = this;
-    connect(detailWindow, &QObject::destroyed, this, [self, windowKey]() {
-        if (self) {
-            self->untrack_window(windowKey);
-        }
-    });
-
-    connect_dialog_close(detailDialog, detailWindow);
-    show_managed_window(detailWindow, listMdiSubWindow_, QPoint(60, 60));
-}
-
-void PartyTypeController::onRevertVersion(const refdata::domain::party_type& type) {
-    BOOST_LOG_SEV(lg(), info) << "Reverting party type to version: " << type.version;
-
-    // Open detail dialog with the old version data for editing
-    auto* detailDialog = new PartyTypeDetailDialog(mainWindow_);
-    if (changeReasonCache_)
-        detailDialog->setChangeReasonCache(changeReasonCache_);
-    detailDialog->setClientManager(clientManager_);
-    detailDialog->setUsername(username_.toStdString());
-    detailDialog->setType(type);
-    detailDialog->setCreateMode(false);
-
-    connect(detailDialog,
-            &PartyTypeDetailDialog::statusMessage,
-            this,
-            &PartyTypeController::statusMessage);
-    connect(detailDialog,
-            &PartyTypeDetailDialog::errorMessage,
-            this,
-            &PartyTypeController::errorMessage);
-    connect(detailDialog,
-            &PartyTypeDetailDialog::typeSaved,
-            this,
-            [self = QPointer<PartyTypeController>(this)](const QString& code) {
-                if (!self)
-                    return;
-                BOOST_LOG_SEV(lg(), info) << "Party Type reverted: " << code.toStdString();
-                emit self->statusMessage(
-                    QString("Party Type '%1' reverted successfully").arg(code));
-                self->handleEntitySaved();
-            });
-
-    auto* detailWindow = new DetachableMdiSubWindow(mainWindow_);
-    detailWindow->setAttribute(Qt::WA_DeleteOnClose);
-    detailWindow->setWidget(detailDialog);
-    detailWindow->setWindowTitle(
-        QString("Revert Party Type: %1").arg(QString::fromStdString(type.code)));
-    detailWindow->setWindowIcon(IconUtils::createRecoloredIcon(Icon::ArrowRotateCounterclockwise,
-                                                               IconUtils::DefaultIconColor));
-
-    register_detachable_window(detailWindow);
-
-    connect_dialog_close(detailDialog, detailWindow);
-    show_managed_window(detailWindow, listMdiSubWindow_);
-}
-
-EntityListMdiWindow* PartyTypeController::listWindow() const {
-    return listWindow_;
-}
-
-}
