+++ projects/ores.refdata/modeling/ores.refdata.party_type.org
@@ -170,6 +170,7 @@ drop function if exists ores_refdata_validate_party_type_fn;
 :item_var:               type
 :key_field:              code
 :has_uuid_primary_key:   false
+:has_change_reason_cache: true
 :get_request_class:      refdata::messaging::get_party_types_request
 :get_response_class:     refdata::messaging::get_party_types_response
 :get_message_type:       get_party_types_request
